feat: reorganize code and expose new `TypedUseQueryOptions`
