+++ src/index.ts
@@ -1,5 +1,5 @@
 export { createQueryKeyStore } from './create-query-key-store';
 export { createQueryKeys } from './create-query-keys';
-export { createMutationKeys } from './create-mutation-keys';
 export { mergeQueryKeys } from './merge-query-keys';
-export type { inferQueryKeys, inferQueryKeyStore } from './types';
+
+export type { TypedUseQueryOptions, inferQueryKeyStore, inferQueryKeys } from './utility-types';
