+++ src/merge-query-keys.ts
@@ -1,13 +1,22 @@
-import { omitPrototype } from './internals';
-import { AnyMutationKeyFactoryResult, AnyQueryKeyFactoryResult, StoreFromMergedQueryKeys } from './types';
+import type { AnyQueryKeyFactoryResult } from './create-query-keys.types';
+import { type Add, omitPrototype } from './internals';
 
-export function mergeQueryKeys<
-  QueryKeyFactoryResults extends Array<AnyQueryKeyFactoryResult | AnyMutationKeyFactoryResult>,
->(...schemas: QueryKeyFactoryResults): StoreFromMergedQueryKeys<QueryKeyFactoryResults> {
+type StoreFromMergedQueryKeys<
+  QueryKeyFactoryResults extends AnyQueryKeyFactoryResult[],
+  CurrentIndex extends number = 0,
+> = QueryKeyFactoryResults[CurrentIndex] extends null | undefined
+  ? {}
+  : {
+      [P in QueryKeyFactoryResults[CurrentIndex]['_def'][0]]: QueryKeyFactoryResults[CurrentIndex];
+    } & StoreFromMergedQueryKeys<QueryKeyFactoryResults, Add<CurrentIndex, 1>>;
+
+export function mergeQueryKeys<QueryKeyFactoryResults extends AnyQueryKeyFactoryResult[]>(
+  ...schemas: QueryKeyFactoryResults
+): StoreFromMergedQueryKeys<QueryKeyFactoryResults> {
   const store = schemas.reduce((storeMap, current) => {
     const [storeKey] = current._def;
 
-    storeMap.set(storeKey, { ...storeMap.get(storeKey), ...current });
+    storeMap.set(storeKey, current);
     return storeMap;
   }, new Map());
